# [RFC] enhance the fmoe heuristic rule for block_n selection