The current implementation of sub_X_cnt doesn’t account for cases where most experts are skipped. For example, when token_cnt = 1, ideally we should select the smallest block_n size to launch more parallel CTAs. However, the existing heuristic chooses 512 (the largest available size).

The root cause is that although the sorting kernel skips padding for empty experts, the length of sorted_expert_ids remains unchanged. In this change, we propose selecting block_n based on an estimated value derived from token_cnt. While we could directly retrieve the concrete token count from a tensor value like num_valid_ids[0], this approach might involve additional D2H operations and potentially break the CUDA graph.

With this adjustment, we’ve observed around a 2x speed-up (135us to 77us)for shapes with small M values.

baseline
```
[aiter] hipModuleLoad: /home/sijiac/oss/aiter/hsa/fmoe/silu/fmoe_fp8_g1u1_subGU_512_silu_tkw1.co GetFunction: fmoe_fp8_g1u1_subGU_512_silu_tkw1 Success
[BW  ] token=1, quant=fp8quant, model_dim=5120, inter_dim=1024, E=128, shared_E=0, topk=1, dtype: torch.bfloat16, asm_bandwidth: 13.57TB/s
[aiter] [perf] use_g1u1=True token=1, quant=fp8quant, model_dim=5120, inter_dim=1024, E=128, shared_E=0, topk=1, dtype: torch.bfloat16, torch_avg: 27963.29 us, asm_avg: 135.19 us ...... uplift: 20583.7%[checkAllclose atol=100 rtol=0.01 failed!]
    a    : torch.Size([1, 5120])
           tensor([ -848.,  -736.,  1064.,   163.,   392., -1064.,   183.,  1360.],
       device='cuda:0', dtype=torch.bfloat16)
    b    : torch.Size([1, 5120])
           tensor([-968., -860.,  944.,  316.,  274., -944.,   44., 1496.],
       device='cuda:0', dtype=torch.bfloat16)
    delta:
           tensor([120., 124., 120., 153., 118., 120., 139., 136.], device='cuda:0',
       dtype=torch.bfloat16)
[aiter] -->max abs delta:216.0, delta details: 6.2% (320 of 5120) elements
```

w/ the change
```
[BW  ] token=1, quant=fp8quant, model_dim=5120, inter_dim=1024, E=128, shared_E=0, topk=1, dtype: torch.bfloat16, asm_bandwidth: 23.66TB/s
[aiter] [perf] use_g1u1=True token=1, quant=fp8quant, model_dim=5120, inter_dim=1024, E=128, shared_E=0, topk=1, dtype: torch.bfloat16, torch_avg: 27336.96 us, asm_avg: 77.54 us ...... uplift: 35157.3%[checkAllclose atol=100 rtol=0.01 failed!]
    a    : torch.Size([1, 5120])
           tensor([ 636., -274., -476.,  236., 1568.,  198., -812., 1000.],
       device='cuda:0', dtype=torch.bfloat16)
    b    : torch.Size([1, 5120])
           tensor([ 748., -154., -620.,  348., 1448.,  318., -960., 1128.],
       device='cuda:0', dtype=torch.bfloat16)
    delta:
           tensor([112., 120., 144., 112., 120., 120., 148., 128.], device='cuda:0',
       dtype=torch.bfloat16)
[aiter] -->max abs delta:208.0, delta details: 6.8% (347 of 5120) elements
```